// portable-deployment/source/src/types.ts
export type StatusOverride = 'none' | 'normal' | 'error'
export type StatusResourceId =
  | 'androidClient'
  | 'windowsClient'
  | 'iosClient'
  | 'macosClient'
  | 'textTranslation'
  | 'cnVoice'
  | 'krVoice'
  | 'illustration'
  | 'announcement'
export type StatusResource = {
  id: StatusResourceId
  resourceVersion: string
  resourceUpdatedAt: string
  officialVersion: string
  officialUpdatedAt: string
  forcedStatus: StatusOverride
}
export type StatusResourceDefinition = {
  id: StatusResourceId
  label: string
}
export type StatusData = { resources: StatusResource[] }